src/main/java/com/jmms/api/domain/jwt/LoginInfo.java
package com.jmms.api.domain.jwt;

public class LoginInfo {

    private String username;
    private String token;
    private Long expiration;

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public Long getExpiration() {
        return expiration;
    }

    public void setExpiration(Long expiration) {
        this.expiration = expiration;
    }

}
